backend/requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
dnspython==2.3.0
docopt==0.6.2
et-xmlfile==1.1.0
Flask==2.2.2
fonttools==4.38.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.1
networkx==2.8.7
numpy==1.23.4
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pipreqs==0.4.11
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
requests==2.28.1
six==1.16.0
soupsieve==2.4.1
urllib3==1.26.12
Werkzeug==2.2.2
yarg==0.1.9
selenium
zipcodes
flask_cors
python-dotenv
